Begin overhaul of t/op/goto.t
Of 134 tests to begin with, 25 were directly affected by changing a
deprecation warning into a fatal exception.  These 25 have been
temporarily commented out (with '##') so that we can decide how to
proceed.  9 of the 109 remaining unit tests are failing; many others are
emitting warnings.

Other test files with failures:

    comp/package_block.t
    lib/croak.t
    porting/deprecation.t
    porting/diag.t
    uni/labels.t

Mark tests which failed with 2 '#' signs