Comment out goto_construct tests
These tests are no longer appropriate given that this deprecation is
being upgraded to an exception.  For the time being, I'm simply
commenting out the tests rather than deleting them, in the hope that we
can replace them with tests of some other deprecation-related warning
(which I have yet to locate).